move bsql statements using legacy back to abstract handler to avoid polluting repository
